requirements.txt
discord.py==1.7.3
requests==2.25.1
python-dotenv==0.17.1
aiohttp==3.7.4.post0